Refactor tracers (required changes from Cloud)
